lib: use core rather than std
`core::fmt` also exports all entities relevant to us, but unlike
`std::fmt` it is available in nostd environments. Likewise, `char` may
be referred to as either `std::char` or `core::char`. This provides the
opportunity to eliminate all our dependencies on `std`, making the crate
usable in nostd environments.